feat: Affiliation and organisation restructure
- Split affiliation and organisation class files
- Added a few from_row and sql_map functions to classes
- Added organisation_info
- Added a related integration test
- Moved the test config out of the git tree

+++ src/route/admin/organisation/affiliation.rs
@@ -8,7 +8,7 @@ use crate::session::UserSession;
 pub fn affiliation_list(
     session: UserSession,
     user_id: Option<u64>,
-    organisation_id: Option<u64>,
+    organisation_id: Option<u32>,
 ) -> Result<Json<Vec<Affiliation>>, Error> {
     let conn = &mut crate::utils::db::get_db_conn()?;
     if !session.right.right_organisation_read {
@@ -23,7 +23,7 @@ pub fn affiliation_list(
 pub fn affiliation_info(
     session: UserSession,
     user_id: u64,
-    organisation_id: u64,
+    organisation_id: u32,
 ) -> Result<Json<Option<Affiliation>>, Error> {
     let conn = &mut crate::utils::db::get_db_conn()?;
     if !session.right.right_organisation_read {
@@ -35,7 +35,7 @@ pub fn affiliation_info(
 }
 
 #[rocket::head("/admin/affiliation_create?<user_id>&<organisation_id>")]
-pub fn affiliation_create(session: UserSession, user_id: u64, organisation_id: u64) -> Result<(), Error> {
+pub fn affiliation_create(session: UserSession, user_id: u64, organisation_id: u32) -> Result<(), Error> {
     let conn = &mut crate::utils::db::get_db_conn()?;
     if !session.right.right_organisation_write {
         return Err(Error::RightOrganisationMissing);
@@ -53,7 +53,7 @@ pub fn affiliation_create(session: UserSession, user_id: u64, organisation_id: u
 pub fn affiliation_edit(
     session: UserSession,
     user_id: u64,
-    organisation_id: u64,
+    organisation_id: u32,
     affiliation: Json<Affiliation>,
 ) -> Result<(), Error> {
     let conn = &mut crate::utils::db::get_db_conn()?;
@@ -66,7 +66,7 @@ pub fn affiliation_edit(
 }
 
 #[rocket::head("/admin/affiliation_delete?<user_id>&<organisation_id>")]
-pub fn affiliation_delete(session: UserSession, user_id: u64, organisation_id: u64) -> Result<(), Error> {
+pub fn affiliation_delete(session: UserSession, user_id: u64, organisation_id: u32) -> Result<(), Error> {
     let conn = &mut crate::utils::db::get_db_conn()?;
     if !session.right.right_organisation_write {
         return Err(Error::RightOrganisationMissing);
